👽 Set `disabled` HTML attributes to `null` when falsy
A [boolean HTML attribute][1] is one that is `true` if it's present,
*regardless* of value, so in:

```html
<input type="checkbox" disabled="false" />
```

...the value of the `disabled` attribute is `true`, because the attribute
is present.

`vue@2` used to coerce falsy boolean HTML attributes such that:

```vue
<input type="checkbox" :disabled="false" />
```

Would render as:

```vue
<input type="checkbox" />
```

However, `vue@3` [removes][2] this coercion, so *now* the above would
render as:

```vue
<input type="checkbox" disabled="false" />
```

...which we saw above means that `disabled` is `true`.

The migration docs say:

> To remove the attribute, use `null` or `undefined`.

This change updates our use of `disabled` to fall back to `null`,
which correctly removes them.

[1]: https://developer.mozilla.org/en-US/docs/Glossary/Boolean/HTML
[2]: https://v3-migration.vuejs.org/breaking-changes/attribute-coercion.html